[Move `@kbn/config-schema` to server] `unified_search`
