+++ src/plugins/unified_search/public/index.ts
@@ -7,7 +7,7 @@
  */
 
 import { PluginInitializerContext } from '@kbn/core/public';
-import { ConfigSchema } from '../config';
+import type { ConfigSchema } from '../server/config';
 export type { IndexPatternSelectProps } from './index_pattern_select';
 export type { QueryStringInputProps } from './query_string_input';
 export { QueryStringInput } from './query_string_input';
