+++ src/plugins/unified_search/server/autocomplete/autocomplete_service.ts
@@ -10,7 +10,7 @@ import moment from 'moment';
 import { clone } from 'lodash';
 import { CoreSetup, Plugin, PluginInitializerContext } from '@kbn/core/server';
 import { registerRoutes } from './routes';
-import { ConfigSchema } from '../../config';
+import { ConfigSchema } from '../config';
 
 export class AutocompleteService implements Plugin<void> {
   private valueSuggestionsEnabled: boolean = true;
