+++ src/plugins/unified_search/server/autocomplete/routes.ts
@@ -9,7 +9,7 @@
 import { Observable } from 'rxjs';
 import { CoreSetup } from '@kbn/core/server';
 import { registerValueSuggestionsRoute } from './value_suggestions_route';
-import { ConfigSchema } from '../../config';
+import { ConfigSchema } from '../config';
 
 export function registerRoutes({ http }: CoreSetup, config$: Observable<ConfigSchema>): void {
   const router = http.createRouter();
