+++ src/plugins/unified_search/server/autocomplete/terms_agg.test.ts
@@ -8,7 +8,7 @@
 
 import { coreMock } from '@kbn/core/server/mocks';
 import { ElasticsearchClient, SavedObjectsClientContract } from '@kbn/core/server';
-import { ConfigSchema } from '../../config';
+import { ConfigSchema } from '../config';
 import type { DeeplyMockedKeys } from '@kbn/utility-types-jest';
 import type { DataViewField, FieldSpec } from '@kbn/data-views-plugin/common';
 import { termsAggSuggestions } from './terms_agg';
