+++ src/plugins/unified_search/server/autocomplete/terms_agg.ts
@@ -11,7 +11,7 @@ import { ElasticsearchClient, SavedObjectsClientContract } from '@kbn/core/serve
 import type * as estypes from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import { getFieldSubtypeNested } from '@kbn/data-plugin/common';
 import type { FieldSpec } from '@kbn/data-views-plugin/common';
-import { ConfigSchema } from '../../config';
+import { ConfigSchema } from '../config';
 import { findIndexPatternById, getFieldByName } from '../data_views';
 
 export async function termsAggSuggestions(
