+++ src/plugins/unified_search/server/autocomplete/terms_enum.test.ts
@@ -9,7 +9,7 @@
 import { termsEnumSuggestions } from './terms_enum';
 import { coreMock } from '@kbn/core/server/mocks';
 import { ElasticsearchClient, SavedObjectsClientContract } from '@kbn/core/server';
-import { ConfigSchema } from '../../config';
+import { ConfigSchema } from '../config';
 import type { DeeplyMockedKeys } from '@kbn/utility-types-jest';
 import { TermsEnumResponse } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import type { DataViewField } from '@kbn/data-views-plugin/common';
