+++ src/plugins/unified_search/server/autocomplete/terms_enum.ts
@@ -10,7 +10,7 @@ import { ElasticsearchClient, SavedObjectsClientContract } from '@kbn/core/serve
 import type * as estypes from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import type { FieldSpec } from '@kbn/data-views-plugin/common';
 import { findIndexPatternById, getFieldByName } from '../data_views';
-import { ConfigSchema } from '../../config';
+import { ConfigSchema } from '../config';
 
 export async function termsEnumSuggestions(
   config: ConfigSchema,
