+++ src/plugins/unified_search/server/autocomplete/value_suggestions_route.ts
@@ -11,7 +11,7 @@ import { IRouter } from '@kbn/core/server';
 import { firstValueFrom, Observable } from 'rxjs';
 import { getRequestAbortedSignal } from '@kbn/data-plugin/server';
 import { getKbnServerError, reportServerError } from '@kbn/kibana-utils-plugin/server';
-import type { ConfigSchema } from '../../config';
+import type { ConfigSchema } from '../config';
 import { termsEnumSuggestions } from './terms_enum';
 import { termsAggSuggestions } from './terms_agg';
