+++ x-pack/plugins/alerting/server/routes/suggestions/values_suggestion_alerts.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { httpServerMock, httpServiceMock } from '@kbn/core-http-server-mocks';
-import type { ConfigSchema } from '@kbn/unified-search-plugin/config';
+import type { ConfigSchema } from '@kbn/unified-search-plugin/server/config';
 import { dataPluginMock } from '@kbn/unified-search-plugin/server/mocks';
 import { termsAggSuggestions } from '@kbn/unified-search-plugin/server/autocomplete/terms_agg';
 import { Observable } from 'rxjs';
