+++ x-pack/plugins/security_solution/server/endpoint/routes/suggestions/index.test.ts
@@ -18,7 +18,7 @@ import type {
   RequestHandlerContext,
   SavedObjectsClientContract,
 } from '@kbn/core/server';
-import type { ConfigSchema } from '@kbn/unified-search-plugin/config';
+import type { ConfigSchema } from '@kbn/unified-search-plugin/server/config';
 import type { Observable } from 'rxjs';
 import { dataPluginMock } from '@kbn/unified-search-plugin/server/mocks';
 import { termsEnumSuggestions } from '@kbn/unified-search-plugin/server/autocomplete/terms_enum';
